requirements.txt
pandas==2.0.2
# dask==2023.6.1
pymongo==4.4.1
qdrant-client==1.3.1
qdrant-txtai==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
sanic>=23.3.0
openai==0.27.4
python-dotenv==1.0.0
python-dateutil~=2.8.2
mysql==0.0.3
redis~=4.6.0
aiohttp~=3.8.5
setuptools==68.0.0
lunardate~=0.2.0